Fix navigation issues on Windows (#52)
- Fix problems navigating to errors and to definitions, references etc. when the compiler is running on Windows.

When the compiler is running on Windows, the data it sends to the VSCE includes carriage returns, which are being interpreted as part of source file URIs. .NET has no way to globally change the line ending emitted by `System.IO.TextWriter.WriteLine()`. However, in the VSCE, all data received from the compiler passes through a single point, so it's cleaner to filter out carriage returns out on this end.